Minor changes to the user list
