+++ src/screens/Social/FriendListScreen.tsx
@@ -13,6 +13,9 @@ import {useDatabaseData} from '@context/global/DatabaseDataContext';
 import {UserArray} from '@src/types/database';
 import UserListComponent from '@components/Social/UserListComponent';
 import useProfileList from '@hooks/useProfileList';
+import {useFocusEffect} from '@react-navigation/native';
+import {arrayItemsAreEqual} from '@libs/Utils';
+import {isNonEmptyArray} from '@libs/Validation';
 
 interface State {
   searching: boolean;
@@ -45,7 +48,7 @@ type FriendListScreenProps = {
 
 const FriendListScreen = (props: FriendListScreenProps) => {
   const {setIndex} = props;
-  const {userData} = useDatabaseData();
+  const {userData, refetch} = useDatabaseData();
   const friendListInputRef = useRef<SearchWindowRef>(null);
   const [state, dispatch] = useReducer(reducer, initialState);
   const {profileList} = useProfileList(state.friends);
@@ -98,6 +101,19 @@ const FriendListScreen = (props: FriendListScreenProps) => {
     dispatch({type: 'SET_FRIENDS', payload: friendsArray});
     dispatch({type: 'SET_FRIENDS_TO_DISPLAY', payload: friendsArray});
   }, [userData]);
+
+  useFocusEffect(
+    React.useCallback(() => {
+      try {
+        refetch(['userData']);
+      } catch (error: any) {
+        Alert.alert(
+          'Failed to contact the database',
+          'Could not update user data:' + error.message,
+        );
+      }
+    }, []),
+  );
 
   return (
     <View style={styles.mainContainer}>
